feat(sidebar): Fase 5 propagação Financeiro — 11 telas + extract FinanceiroSubNav shared (ADR 0180)
ADR 0180 Fase 5 propagação do pattern aprovado por Wagner em PR #1365 (ghost
tabs + ⋯ Mais + primary "+ Novo título") pras 11 telas restantes do Financeiro.

Etapa 1 — extração shared:
  FinanceiroSubNav era inline em Unificado/Index.tsx (linhas 764-809).
  Extraído pra resources/js/Pages/Financeiro/_shared/FinanceiroSubNav.tsx
  com exports `FinanceiroSubNav` (default) + tipos `FinanceiroSubNavProps`.
  Unificado/Index.tsx agora importa o componente shared (-46 LOC).

Etapa 2 — propagação 11 telas:
  ContasReceber  · active="contas-receber"
  ContasPagar    · active="contas-pagar"
  Fluxo          · active="fluxo"
  Cobranca       · active="cobranca"
  Caixa          · active="caixa"
  Conciliacao    · active="conciliacao"
  Dre            · active="dre"
  Relatorios     · active="relatorios"
  ContasBancarias· active="contas-bancarias"
  PlanoContas    · active="plano-contas"
  Categorias     · active="categorias"

Pattern uniforme: insere <FinanceiroSubNav active="..."/> dentro do
`<div className="os-page-h-r fin-page-h-r">` do `<header className="os-page-h">`
de cada tela. Telas que já tinham `os-page-h-r` ganham SubNav antes dos botões
existentes (Cobranca/Dre/Relatorios/ContasBancarias/Categorias). Telas que só
tinham `os-page-h-l` ganham div `os-page-h-r` nova com o SubNav
(ContasReceber/ContasPagar/Fluxo/Caixa/Conciliacao/PlanoContas).

Sem props `extraOverflowItems` nem `hidePrimary` por enquanto — refinements
caso-a-caso depois (Wagner). Botões "Novo X" próprios das telas preservados
sem mexer.

Refs: ADR 0180 (sidebar v3 + PageHeaderTabs canon) · PR #1365 piloto
