+++ resources/js/Pages/Financeiro/PlanoContas/Index.tsx
@@ -10,6 +10,7 @@
 import AppShellV2 from '@/Layouts/AppShellV2';
 import { type ReactNode, useMemo, useState } from 'react';
 import { Lock, FileText, Search } from 'lucide-react';
+import FinanceiroSubNav from '@/Pages/Financeiro/_shared/FinanceiroSubNav';
 
 interface PlanoConta {
   id: number;
@@ -69,6 +70,10 @@ function FinanceiroPlanoContas({ planos, stats }: Props) {
         <div className="os-page-h-l fin-page-h-l">
           <h1>Plano de Contas <span className="fin-hero-title-sub">· Estrutura contábil BR</span></h1>
           <p>{stats.total} contas hierárquicas (Receita Federal/DCASP) — Eliana classifica lançamentos pelo plano</p>
+        </div>
+        <div className="os-page-h-r fin-page-h-r">
+          {/* ADR 0180 Fase 5 propagação — ghost tabs Financeiro + primary `+ Novo título` */}
+          <FinanceiroSubNav active="plano-contas" />
         </div>
       </header>
